requirements.txt
aiohttp==3.7.4.post0
aiohttp-proxy==0.1.2
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.9.3
certifi==2021.10.8
chardet==3.0.4
discord.py==1.7.3
dnspython==2.1.0
idna==2.8
multidict==6.0.2
pymongo==3.11.4
pysaucenao==1.6.1
python-dotenv==0.17.1
requests==2.21.0
soupsieve==2.3.2
typing_extensions==4.1.1
urllib3==1.24.3
yarl==1.7.2
